python-recommendation-service/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pandas==2.3.0
numpy==2.1.2
scikit-learn==1.5.2
scipy==1.14.1
psycopg2-binary==2.9.10